Implement type safety that now exists.
